Added Singleton Design Pattern - Java
